refactor: new inspect module
